fix(Tca\ContentType): fix multiple issues when handling content type rows

+++ Classes/ExtConfigHandler/Table/ContentTypeApplier.php
@@ -189,25 +189,25 @@ class ContentTypeApplier extends AbstractExtConfigApplier
             return;
         }
         
-        // Clean up after ourselves if the cType of a row previously HAD an extension
-        // but now has none, OR the CType was changed, we drop everything
-        $hasExtensionTable = ContentTypeUtil::hasExtensionTable($cType);
-        $oldCType = $this->resolveColumnValue('CType', ['uid' => $event->getId()]);
-        if (! $hasExtensionTable || (is_string($oldCType) && $oldCType !== $cType && $hasExtensionTable)) {
+        // If there is no extension table for the new cType we drop everything
+        // and let the data handler go on by itself.
+        if (! ContentTypeUtil::hasExtensionTable($cType)) {
             // There are no extension tables -> skip
             if (empty(ContentTypeUtil::getTableMap())) {
                 return;
             }
             
             $row = ContentTypeUtil::removeAllExtensionColumns($row);
-            $row['ct_child'] = '';
+            $row['ct_child'] = '0';
             $event->setRow($row);
             
-            if ($hasExtensionTable) {
-                $this->repository->deleteChildRow($oldCType, $event->getId());
-            }
-            
             return;
+        }
+        
+        // When the cType is being changed we try to drop the old row
+        $oldCType = $this->resolveColumnValue('CType', ['uid' => $event->getId()]);
+        if (is_string($oldCType) && $oldCType !== $cType && ContentTypeUtil::hasExtensionTable($oldCType)) {
+            $this->repository->deleteChildRow($oldCType, $event->getId());
         }
         
         $childRow = ContentTypeUtil::extractChildFromParent($row, $cType);
